chatiot/client/requirements.txt
backports.statistics==0.1.0
loguru==0.7.2
numpy==2.1.0
opencv_python==4.10.0.84
Pillow==10.4.0
pvporcupine==3.0.2
PyAudio==0.2.14
pydantic==2.8.2
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
sentence_transformers==3.0.1
streamlit==1.35.0
streamlit_chatbox==1.1.12.post2
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
websocket_client==1.8.0
